+++ lib/model/book.dart
@@ -1,6 +1,6 @@
 class Book {
   Book(this.id, this.imageUrl);
 
-  final BigInt id;
+  final int id;
   final String imageUrl;
 }
